% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{has_tiers}
\alias{has_tiers}
\title{Check that tiers are in a textgrid}
\usage{
has_tiers(textgrid, tiers)
}
\arguments{
\item{textgrid}{a textgrid object}

\item{tiers}{names of tiers to check for in the textgrid}
}
\value{
TRUE if the test passes
}
\description{
Check that tiers are in a textgrid
}

