% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalTier-accessors.R
\name{IntervalTier-accessors}
\alias{IntervalTier-accessors}
\alias{intervalEndTimes}
\alias{intervalLabels}
\alias{intervalStartTimes}
\title{Access the slots of IntervalTier objects.}
\usage{
intervalStartTimes(tier)

intervalEndTimes(tier)

intervalLabels(tier)
}
\arguments{
\item{tier}{An \code{\link[=IntervalTier-class]{IntervalTier}} object.}
}
\description{
Functions for accessing the slots of an \code{\link[=IntervalTier-class]{IntervalTier}}
object.
}
\seealso{
\code{\link{IntervalTier-class}},
  \code{\link{IntervalTier-constructor}}, \code{\link{Tier-accessors}}
}

