% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is_text_tier}
\alias{is_text_tier}
\title{Check that the named tiers are text/point tiers}
\usage{
is_text_tier(textgrid, tiers)
}
\arguments{
\item{textgrid}{a textgrid object}

\item{tiers}{names of tiers in the textgrid}
}
\value{
TRUE if the test passes
}
\description{
Check that the named tiers are text/point tiers
}

