% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointTier-accessors.R
\name{PointTier-accessors}
\alias{PointTier-accessors}
\alias{pointLabels}
\alias{pointTimes}
\title{Access the slots of PointTier objects.}
\usage{
pointTimes(tier)

pointLabels(tier)
}
\arguments{
\item{tier}{A \code{\link[=PointTier-class]{PointTier}} object.}
}
\description{
Functions for accessing the slots of a \code{\link[=PointTier-class]{PointTier}}
object.
}
\seealso{
\code{\link{PointTier-class}},
  \code{\link{PointTier-constructor}}, \code{\link{Tier-accessors}}
}

