% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointTier-class.R
\docType{class}
\name{PointTier-class}
\alias{PointTier-class}
\title{PointTier S4 class for Praat TextGrids.}
\description{
The \code{PointTier} class extends the \code{\link[=Tier-class]{Tier}} class.
A \code{PointTier} object describes a sequence of labeled points in time.
A point's label is typically the annotation of some event in waveform data
(e.g., the onset of voicing in speech data).
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string, the name of the Tier.}

\item{\code{number}}{An integer, the number of the Tier within the TextGrid.}

\item{\code{times}}{A numeric vector, the times of the points in the PointTier.}

\item{\code{labels}}{A character vector, the labels of the points in the PointTier.}
}}
\seealso{
\code{\link{PointTier-constructor}},
  \code{\link{PointTier-accessors}}, \code{\link{TextGrid-class}},
  \code{\link{Tier-class}}
}

