% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointTier-constructor.R
\docType{methods}
\name{PointTier-constructor}
\alias{PointTier}
\alias{PointTier,character-method}
\alias{PointTier-constructor}
\title{Create an instance of the PointTier class.}
\usage{
PointTier(praatText, ...)

\S4method{PointTier}{character}(praatText)
}
\arguments{
\item{praatText}{A character vector, the lines of text from a
\code{.TextGrid} file that define a PointTier.}

\item{...}{optional arguments for multiple dispatch (in development).}
}
\value{
A \code{\link[=PointTier-class]{PointTier}} object. Values for the
  \code{tierName}, \code{tierNumber}, \code{times}, and \code{labels}
  slots are parsed automatically from the \code{praatText}.
}
\description{
An S4 generic and S4 methods for creating an \code{\link[=PointTier-class]{PointTier}}
object.
}
\seealso{
\code{\link{PointTier-class}}, \code{\link{PointTier-accessors}}
}

