% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextGrid-accessors.R
\name{TextGrid-accessors}
\alias{TextGrid-accessors}
\alias{textGridEndTime}
\alias{textGridStartTime}
\title{Access the slots of TextGrid objects.}
\usage{
textGridStartTime(textGrid)

textGridEndTime(textGrid)
}
\arguments{
\item{textGrid}{An \code{\link[=TextGrid-class]{TextGrid}} object.}
}
\description{
Functions for accessing the slots of a \code{\link[=TextGrid-class]{TextGrid}}
object.
}
\seealso{
\code{\link{TextGrid-class}},
  \code{\link{TextGrid-constructor}}
}

