% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{textgRid-as.data.frame}
\alias{as.data.frame.IntervalTier}
\alias{as.data.frame.PointTier}
\alias{as.data.frame.TextGrid}
\alias{textgRid-as.data.frame}
\title{Coerce to a data.frame.}
\usage{
\method{as.data.frame}{IntervalTier}(x, stringsAsFactors = FALSE)

\method{as.data.frame}{PointTier}(x, stringsAsFactors = FALSE)

\method{as.data.frame}{TextGrid}(x, stringsAsFactors = FALSE)
}
\arguments{
\item{x}{An \code{IntervalTier}, \code{PointTier}, or \code{TextGrid} object.}

\item{stringsAsFactors}{A logical, default is \code{FALSE}.}
}
\value{
A \code{data.frame} object whose rows represent the contentful
  intervals or points within \code{object}, and whose columns are:
  \code{TierNumber}, \code{TierName}, \code{TierType},
  \code{Index}, \code{StartTime}, \code{EndTime}, \code{Label}.
}
\description{
Coerce an \code{IntervalTier}, \code{PointTier}, or \code{TextGrid} object
to a \code{data.frame}.
}
\details{
\code{IntervalTier}s and \code{PointTier}s are coerced by passing them to
\code{findIntervals()} and \code{findPoints()}, respectively. Only intervals
and points that have contentful, non-empty labels are returned after coercion.
}

