% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextGrid-class.R
\docType{class}
\name{TextGrid-class}
\alias{TextGrid-class}
\title{TextGrid S4 class for Praat TextGrids.}
\description{
The \code{TextGrid} class extends the \code{\link{list}} class.
A \code{TextGrid} object is essentially a list of
\code{\link[=IntervalTier-class]{IntervalTier}} and
\code{\link[=PointTier-class]{PointTier}} objects.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list of
\code{\link[=IntervalTier-class]{IntervalTier}} and
\code{\link[=PointTier-class]{PointTier}} objects.}

\item{\code{startTime}}{A numeric, the start time of the TextGrid.}

\item{\code{endTime}}{A numeric, the end time of the TextGrid.}
}}
\seealso{
\code{\link{TextGrid-constructor}},
  \code{\link{TextGrid-accessors}}, \code{\link{IntervalTier-class}},
  \code{\link{PointTier-class}}
}

