% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextGrid-constructor.R
\docType{methods}
\name{TextGrid-constructor}
\alias{TextGrid-constructor}
\alias{TextGrid}
\alias{TextGrid,character-method}
\title{Create an instance of the TextGrid class.}
\usage{
TextGrid(textGrid, ...)

\S4method{TextGrid}{character}(textGrid, encoding = NULL)
}
\arguments{
\item{textGrid}{A character vector}

\item{...}{optional arguments for multiple dispatch (in development).}

\item{encoding}{The character encoding of the .TextGrid file. If \code{NULL},
then the encoding of the file is guessed using \code{\link[readr]{guess_encoding}}.
Plausible encodings that might be used by Praat are \code{"ASCII"},
\code{"UTF-8"}, or \code{"UTF-16BE"} (if non-ASCII characters occur in
the TextGrid).}
}
\value{
A \code{\link[=TextGrid-class]{TextGrid}} object.
}
\description{
An S4 generic and S4 methods for creating a \code{\link[=TextGrid-class]{TextGrid}}
object.
}
\section{Details for signature \code{c(textGrid = 'character')}}{

 If \code{textGrid} is a string (i.e., a character vector with
 \code{length(textGrid)=1}), then it is assumed that the \code{textGrid}
 argument is the path to a \code{.TextGrid} file. Otherwise, the
 \code{textGrid} argument is assumed to be a character vector whose
 elements are the lines of some \code{.TextGrid} file.
}

\seealso{
\code{\link{TextGrid-class}}, \code{\link{TextGrid-accessors}}
}
