% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{textgRid-length}
\alias{textgRid-length}
\alias{length.IntervalTier}
\alias{length.PointTier}
\alias{length.IntervalTier}
\alias{length.PointTier}
\title{Length of an IntervalTier or PointTier}
\usage{
\method{length}{IntervalTier}(x)

\method{length}{PointTier}(x)
}
\arguments{
\item{x}{An \code{IntervalTier} or \code{PointTier} object.}
}
\value{
A \code{numeric}. The number of intervals or points on \code{x}.
}
\description{
Get the number of intervals on an \code{IntervalTier} or the number of points
on a \code{PointTier}.
}
\details{
\code{length.IntervalTier()} checks that the vectors stored in the
\code{startTimes}, \code{endTimes}, and \code{labels} slots of the
\code{IntervalTier} object all have the same length. If so, the length
of the \code{labels}-vector is returned; otherwise, \code{NULL} is returned.

\code{length.PointTier()} checks that the vectors stored in the
\code{times} and \code{labels} slots of the \code{PointTier} object both
have the same length. If so, the length of the \code{labels}-vector is
returned; otherwise, \code{NULL} is returned.
}
