% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TextGrid.R
\name{TextGrid}
\alias{TextGrid}
\title{TextGrid class}
\usage{
TextGrid(textGridData, textGridXMin, textGridXMax)
}
\description{
TextGrid class
}
\section{Slots}{

\describe{
\item{\code{textGridXMin}}{}

\item{\code{textGridXMax}}{}
}}

