% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tier-accessors.R
\name{Tier-accessors}
\alias{Tier-accessors}
\alias{tierName}
\alias{tierNumber}
\title{Methods for \code{Tier} objects.}
\usage{
tierName(tier)

tierNumber(tier)
}
\arguments{
\item{tier}{A \code{\link[=Tier-class]{Tier}} object.}
}
\description{
Get the values of slots in a \code{\link[=Tier-class]{Tier}} object.
}
\seealso{
\code{\link{Tier-class}}
}

