% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTextGrid.R
\name{writeTextGrid}
\alias{writeTextGrid}
\title{Write Praat-compatible TextGrid.}
\usage{
writeTextGrid(x, path = NULL, ...)
}
\arguments{
\item{x}{A \code{TextGrid} object to be written.}

\item{path}{Either a character string naming a file to write to, a connection 
open for writing, or \code{NULL} (default) for no output. When writing to 
file or connection, \code{path} is passed on as the \code{con} argument to 
\code{\link{writeLines}}}

\item{...}{Additional arguments passed on to \code{\link{writeLines}} 
when writing to a file or connection.}
}
\value{
A character vector, Each element is one row of the TextGrid file.
}
\description{
Convert a \code{TextGrid} object to a Praat-compatible character string and 
(optionally) write it to a file.
}
\seealso{
\code{\link{TextGrid-class}}
}
