% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalTier-constructor.R
\docType{methods}
\name{IntervalTier-constructor}
\alias{IntervalTier-constructor}
\alias{IntervalTier}
\alias{IntervalTier,character-method}
\alias{IntervalTier,data.frame-method}
\title{Create an instance of the IntervalTier class.}
\usage{
IntervalTier(x, ...)

\S4method{IntervalTier}{character}(x)

\S4method{IntervalTier}{data.frame}(x)
}
\arguments{
\item{...}{optional arguments for multiple dispatch (in development).}

\item{praatText}{A character vector, the lines of text from a
\code{.TextGrid} file that define an IntervalTier.}
}
\value{
A \code{\link[=IntervalTier-class]{IntervalTier}} object. Values for the
  \code{tierName}, \code{tierNumber}, \code{startTimes}, \code{endTimes},
  and \code{labels} slots are parsed automatically from the \code{praatText}.
}
\description{
An S4 generic and S4 methods for creating an \code{\link[=IntervalTier-class]{IntervalTier}}
object.
}
\seealso{
\code{\link{IntervalTier-class}}, \code{\link{IntervalTier-accessors}}
}
