% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgRid.R
\docType{package}
\name{textgRid}
\alias{textgRid}
\alias{textgRid-package}
\title{textgRid: Praat TextGrid Objects in R}
\description{
The software application Praat can be used to annotate
waveform data (e.g., to mark intervals of interest or to label events).
These annotations are stored in a Praat TextGrid object, which consists of
a number of interval tiers and point tiers. An interval tier consists of
sequential (i.e., not overlapping) labeled intervals. A point tier consists
of labeled events that have no duration. The textgRid package provides
S4 classes, generics, and methods for accessing information that is stored
in Praat TextGrid objects.
}
\section{S4 classes}{

\code{\link[=Tier-class]{Tier}},
\code{\link[=IntervalTier-class]{IntervalTier}},
\code{\link[=PointTier-class]{PointTier}},
\code{\link[=TextGrid-class]{TextGrid}}
}

\section{S4 generics and methods}{

\code{\link[=TextGrid-constructor]{TextGrid()} object constructor}
}

\section{Functions}{

\code{\link{findIntervals}},
\code{\link{findPoints}}
}

