% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/TextGrid.R
\docType{methods}
\name{textGridName}
\alias{FormatAsPraatText,TextGrid-method}
\alias{GetIntervalText,TextGrid,ANY,numeric-method}
\alias{GetIntervalXMax,TextGrid,ANY,numeric-method}
\alias{GetIntervalXMin,TextGrid,ANY,numeric-method}
\alias{TextGrid}
\alias{TextGrid,IntervalTier-method}
\alias{TextGrid,character-method}
\alias{TimeSlice,TextGrid,numeric,numeric-method}
\alias{endTime,TextGrid-method}
\alias{size,TextGrid-method}
\alias{startTime,TextGrid-method}
\alias{textGridName}
\alias{textGridName,TextGrid-method}
\alias{timeUnit,TextGrid-method}
\title{TextGrid class}
\usage{
textGridName(.Object)

TextGrid(x, ...)

TextGrid(x, ...)

\S4method{TextGrid}{character}(x)

\S4method{TextGrid}{IntervalTier}(x, ...)

\S4method{textGridName}{TextGrid}(.Object)

\S4method{size}{TextGrid}(.Object)

\S4method{startTime}{TextGrid}(.Object)

\S4method{endTime}{TextGrid}(.Object)

\S4method{timeUnit}{TextGrid}(.Object)

\S4method{FormatAsPraatText}{TextGrid}(x)

\S4method{TimeSlice}{TextGrid,numeric,numeric}(x, sliceFrom, sliceTo, sliceUnit)

\S4method{GetIntervalText}{TextGrid,ANY,numeric}(textGrid, tier, interval)

\S4method{GetIntervalXMin}{TextGrid,ANY,numeric}(textGrid, tier, interval)

\S4method{GetIntervalXMax}{TextGrid,ANY,numeric}(textGrid, tier, interval)
}
\description{
TextGrid class
}
\section{Slots}{

\describe{
\item{\code{textGridName}}{name of the text grid file}

\item{\code{size}}{number of tiers}

\item{\code{startTime}}{start time of text grid}

\item{\code{endTime}}{end time of text grid}

\item{\code{timeUnit}}{measurement unit of the start and end times (seconds,
milliseconds, microseconds, nanoseconds)}
}}

